+++ exercise/kotlin/day13/src/main/kotlin/santamarket/model/offer/TenPercentOffer.kt
@@ -1,0 +1,23 @@
+package santamarket.model.offer
+
+import santamarket.model.Discount
+import santamarket.model.Product
+import santamarket.model.SpecialOfferType
+
+class TenPercentOffer(offerType: SpecialOfferType, product: Product, argument: Double) :
+    Offer(offerType, product, argument) {
+
+    override fun getDiscount(
+        unitPrice: Double,
+        itemsInCart: Double
+    ) = getTenPercentDiscount(unitPrice, itemsInCart)
+
+    private fun getTenPercentDiscount(
+        unitPrice: Double,
+        itemsInCart: Double,
+    ): Discount {
+        val product = this.product
+        return Discount(product, "${this.argument}% off", -getUndiscountedPrice(unitPrice, itemsInCart) * this.argument / 100.0)
+    }
+
+}
